+++ R/stem.R
@@ -149,7 +149,7 @@ bake.step_stem <- function(object, new_data, ...) {
 
   for (i in seq_along(col_names)) {
     stemmed_tokenlist <- tokenlist_apply(new_data[, col_names[i], drop = TRUE],
-                                         stem_fun)
+                                         stem_fun, object$options)
 
     new_data[, col_names[i]] <- tibble(stemmed_tokenlist)
   }